+++ BE/Spring/audisay/src/main/java/com/palja/audisay/domain/recommendation/entity/DemographicsBook.java
@@ -16,5 +16,5 @@ public class DemographicsBook {
 	@Id
 	private String id;
 	private String groupId;
-	private List<Integer> bookList;
+	private List<Long> bookList;
 }
