+++ BE/Spring/audisay/src/main/java/com/palja/audisay/domain/recommendation/entity/SimilarMemberBook.java
@@ -16,5 +16,5 @@ public class SimilarMemberBook {
 	@Id
 	private String id;
 	private Long memberId;
-	private List<Integer> bookList;
+	private List<Long> bookList;
 }
